auth_login_system/auth_system/requirements.txt
alembic==1.18.4
async-timeout==5.0.1
bcrypt==4.0.1
blinker==1.9.0
click==8.3.1
coverage==7.13.4
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.1.0
exceptiongroup==1.3.1
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Cors==4.0.1
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
gunicorn==21.2.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.8.0
pytest==7.4.3
pytest-cov==4.1.0
python-dotenv==1.0.0
redis==7.2.1
rich==13.9.4
SQLAlchemy==2.0.47
tomli==2.4.0
typing_extensions==4.15.0
Werkzeug==2.3.7
wrapt==2.1.1
WTForms==3.2.1
